Cleaned up documentation, corrected auth url, and defaults.
Had left over references to the ueberauth_github strategy documention. Hopefully those are all gone now. I went through the docs better and made sure they were applicable to this project as well. In doing so, I found where some of the configuration wasn't used correctly (uid and scopes).

Tried to remove a few of the loggers and IO.puts in the code. I will probably end up putting in Logger.debug for users of the project in case they need to figure something out (like they didn't specify the right scope ...etc). But won't right now.

This works for an unreleased version of oauth2 that I submitted a PR for. So basically this project won't work for others until that is released.